docs(live-preview): dynamic import modules by import statements
